+++ tests/integ/modin/frame/test_to_snowflake.py
@@ -3,29 +3,106 @@
 #
 
 import logging
+from random import Random
 import re
+import string
 
 import modin.pandas as pd
 import pandas as native_pd
 import pytest
+from modin.config import context as config_context
 
 import snowflake.snowpark.modin.plugin  # noqa: F401
 from tests.integ.modin.utils import VALID_PANDAS_LABELS, VALID_SNOWFLAKE_COLUMN_NAMES
 from tests.integ.utils.sql_counter import SqlCounter, sql_count_checker
+from modin.config import PandasToSnowflakeParquetThresholdBytes
+
+from tests.utils import Utils
+
+
+def _generate_lower_case_table_name_unquoted() -> str:
+    return "snowpark_temp_table_" + "".join(
+        Random().choice(string.ascii_lowercase + string.digits) for _ in range(10)
+    )
+
+
+@pytest.fixture
+def lower_case_table_name_unquoted(session):
+    name = _generate_lower_case_table_name_unquoted()
+    yield name
+    Utils.drop_table(session, name)
+
+
+@pytest.fixture
+def lower_case_table_name_quoted(session):
+    name = '"' + _generate_lower_case_table_name_unquoted() + '"'
+    yield name
+    Utils.drop_table(session, name)
+
+
+@pytest.fixture
+def upper_case_table_name_quoted(session):
+    name = '"' + _generate_lower_case_table_name_unquoted().upper() + '"'
+    yield name
+    Utils.drop_table(session, name)
+
+
+@pytest.fixture
+def upper_case_table_name_with_space_quoted(session):
+    name = '" ' + _generate_lower_case_table_name_unquoted().upper() + ' "'
+    yield name
+    Utils.drop_table(session, name)
+
+
+@pytest.fixture(params=("pandas", "snowflake"), autouse=True)
+def starting_backend(request) -> str:
+    with config_context(Backend=request.param):
+        yield
+
+
+@pytest.fixture(
+    params=[0, int(1e9)], autouse=True, ids=lambda param: f"parquet_threshold_{param}"
+)
+def parquet_threshold(request):
+    with config_context(PandasToSnowflakeParquetThresholdBytes=request.param):
+        yield
+
+
+def to_snowflake_counter(*, df: pd.DataFrame, if_exists: str) -> SqlCounter:
+    if (
+        df.get_backend() == "Pandas"
+        and df.memory_usage(deep=False).sum()
+        > PandasToSnowflakeParquetThresholdBytes.get()
+    ):
+        # In this case we are using a parquet file to load pandas data to
+        # Snowflake.
+        # if if_exists='fail', we issue a query to check whether the table
+        # exists. In any case, we do issue more queries to make the parquet
+        # file, stage it, and load it into a table, but the SqlCounter doesn't
+        # track those queries.
+        query_count = 1 if if_exists == "fail" else 0
+    elif if_exists in ("append", "fail"):
+        query_count = 2
+    else:
+        assert if_exists == "replace"
+        # In this case, we can skip the query that checks whether the table
+        # exists.
+        query_count = 1
+    return SqlCounter(query_count=query_count)
 
 
 @pytest.mark.parametrize("index", [True, False])
 @pytest.mark.parametrize("index_labels", [None, ["my_index"]])
-# one extra query to convert index to native pandas when creating the snowpark pandas dataframe
-@sql_count_checker(query_count=2)
 def test_to_snowflake_index(test_table_name, index, index_labels):
     df = pd.DataFrame(
-        {"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4], name="index")
+        {"a": [1, 2, 3], "b": [4, 5, 6]}, index=native_pd.Index([2, 3, 4], name="index")
     )
 
-    df.to_snowflake(
-        test_table_name, if_exists="replace", index=index, index_label=index_labels
-    )
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(
+            test_table_name, if_exists=if_exists, index=index, index_label=index_labels
+        )
     expected_columns = []
     if index:
         # if index is retained in the result, add it as the first expected column
@@ -38,7 +115,6 @@ def test_to_snowflake_index(test_table_name, index, index_labels):
     verify_columns(test_table_name, expected_columns)
 
 
-@sql_count_checker(query_count=1)
 def test_to_snowflake_multiindex(test_table_name):
     index = native_pd.MultiIndex.from_arrays(
         [[1, 1, 2, 2], ["red", "blue", "red", "blue"]], names=("number", "color")
@@ -47,7 +123,9 @@ def test_to_snowflake_multiindex(test_table_name):
         [[1] * 2, [2] * 2, [3] * 2, [4] * 2], index=index, columns=["a", "b"]
     )
     snow_df = pd.DataFrame(native_df)
-    snow_df.to_snowflake(test_table_name, if_exists="replace", index=True)
+    if_exists = "replace"
+    with to_snowflake_counter(df=snow_df, if_exists=if_exists):
+        snow_df.to_snowflake(test_table_name, if_exists=if_exists, index=True)
     verify_columns(test_table_name, ["number", "color", "a", "b"])
 
     with pytest.raises(
@@ -94,14 +172,17 @@ def test_to_snowflake_if_exists(session, test_table_name):
     df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
 
     # Verify new table is created
-    with SqlCounter(query_count=2):
-        df.to_snowflake(test_table_name, if_exists="fail", index=False)
-        verify_columns(test_table_name, ["a", "b"])
+    if_exists = "fail"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
+    verify_columns(test_table_name, ["a", "b"])
 
-    # Verify attempt to write to existing table fails
+    # Verify attempt to write to existing table fails.
+    # Just one query to find that the table already exists.
+    if_exists = "fail"
     with SqlCounter(query_count=1):
         with pytest.raises(ValueError):
-            df.to_snowflake(test_table_name, if_exists="fail", index=False)
+            df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
 
     # Verify by default attempt to write to existing table fails
     with SqlCounter(query_count=1):
@@ -109,23 +190,26 @@ def test_to_snowflake_if_exists(session, test_table_name):
             df.to_snowflake(test_table_name, index=False)
 
     # Verify existing table is replaced with new data
+    if_exists = "replace"
     df = pd.DataFrame({"a": [1, 2, 3], "c": [4, 5, 6]})
-    with SqlCounter(query_count=2):
-        df.to_snowflake(test_table_name, if_exists="replace", index=False)
-        verify_columns(test_table_name, ["a", "c"])
-        verify_num_rows(session, test_table_name, 3)
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
+    verify_columns(test_table_name, ["a", "c"])
+    verify_num_rows(session, test_table_name, 3)
 
     # Verify data is appended to existing table
-    with SqlCounter(query_count=3):
-        df.to_snowflake(test_table_name, if_exists="append", index=False)
-        verify_columns(test_table_name, ["a", "c"])
-        verify_num_rows(session, test_table_name, 6)
+    if_exists = "append"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
+    verify_columns(test_table_name, ["a", "c"])
+    verify_num_rows(session, test_table_name, 6)
 
     # Verify pd.to_snowflake operates the same
-    with SqlCounter(query_count=3):
-        pd.to_snowflake(df, test_table_name, if_exists="append", index=False)
-        verify_columns(test_table_name, ["a", "c"])
-        verify_num_rows(session, test_table_name, 9)
+    if_exists = "append"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        pd.to_snowflake(df, test_table_name, if_exists=if_exists, index=False)
+    verify_columns(test_table_name, ["a", "c"])
+    verify_num_rows(session, test_table_name, 9)
 
     # Verify invalid 'if_exists' input.
     with SqlCounter(query_count=0):
@@ -134,87 +218,96 @@ def test_to_snowflake_if_exists(session, test_table_name):
 
 
 @pytest.mark.parametrize("index_label", VALID_PANDAS_LABELS)
-@sql_count_checker(query_count=1)
 def test_to_snowflake_index_labels(index_label, test_table_name):
     df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
-    df.to_snowflake(
-        test_table_name, if_exists="replace", index=True, index_label=index_label
-    )
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(
+            test_table_name, if_exists=if_exists, index=True, index_label=index_label
+        )
     verify_columns(test_table_name, [str(index_label), "a", "b"])
 
 
 @pytest.mark.parametrize("col_name", VALID_PANDAS_LABELS)
-@sql_count_checker(query_count=1)
-def test_to_snowflake_column_names_from_panadas(col_name, test_table_name):
+def test_to_snowflake_column_names_from_pandas(col_name, test_table_name):
     df = pd.DataFrame({col_name: [1, 2, 3], "b": [4, 5, 6]})
-    df.to_snowflake(test_table_name, if_exists="replace", index=False)
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists="replace", index=False)
     verify_columns(test_table_name, [str(col_name), "b"])
 
 
 @pytest.mark.parametrize("col_name", VALID_SNOWFLAKE_COLUMN_NAMES)
 @pytest.mark.parametrize("if_exists", ["append", "replace"])
 def test_column_names_with_read_snowflake_and_to_snowflake(
-    col_name, if_exists, session
+    col_name, if_exists, session, test_table_name
 ):
-    with SqlCounter(query_count=6 if if_exists == "append" else 5):
-        # Create a table
-        session.sql(f"create or replace table t1 ({col_name} int)").collect()
-        session.sql("insert into t1 values (1), (2), (3)").collect()
-        data = session.sql(f"select {col_name} from t1").collect()
-        assert len(data) == 3
-        # Capture column names of origing table.
-        expected_columns = session.table("t1").columns
+    # Create a table
+    session.sql(f"create or replace table {test_table_name} ({col_name} int)").collect()
+    session.sql(f"insert into {test_table_name} values (1), (2), (3)").collect()
+    data = session.sql(f"select {col_name} from {test_table_name}").collect()
+    assert len(data) == 3
 
-        df = pd.read_snowflake("t1")
-        df.to_snowflake("t1", if_exists=if_exists, index=False)
-        # Verify column names are not updated here.
-        assert expected_columns == session.table("t1").columns
-        data = session.sql(f"select {col_name} from t1").collect()
-        assert len(data) == (6 if if_exists == "append" else 3)
+    # Capture column names of origing table.
+    expected_columns = session.table(test_table_name).columns
+
+    df = pd.read_snowflake(test_table_name)
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
+    # Verify column names are not updated here.
+    assert expected_columns == session.table(test_table_name).columns
+    data = session.sql(f"select {col_name} from {test_table_name}").collect()
+    assert len(data) == (6 if if_exists == "append" else 3)
 
 
-@sql_count_checker(query_count=1)
 def test_to_snowflake_column_with_quotes(session, test_table_name):
     df = pd.DataFrame({'a"b': [1, 2, 3], 'a""b': [4, 5, 6]})
-    df.to_snowflake(test_table_name, if_exists="replace", index=False)
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index=False)
     verify_columns(test_table_name, ['a"b', 'a""b'])
 
 
 # one extra query to convert index to native pandas when creating the snowpark pandas dataframe
 def test_to_snowflake_index_label_none(test_table_name):
     # no index
-    with SqlCounter(query_count=1):
-        df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
-        df.to_snowflake(test_table_name, if_exists="replace")
-        verify_columns(test_table_name, ["index", "a", "b"])
+    df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]})
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists)
+    verify_columns(test_table_name, ["index", "a", "b"])
 
     # named index
-    with SqlCounter(query_count=2):
-        df = pd.DataFrame(
-            {"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4], name="index")
-        )
-        df.to_snowflake(test_table_name, if_exists="replace", index_label=[None])
-        verify_columns(test_table_name, ["index", "a", "b"])
+    if_exists = "replace"
+    df = pd.DataFrame(
+        {"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4], name="index")
+    )
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index_label=[None])
+    verify_columns(test_table_name, ["index", "a", "b"])
 
     # nameless index
-    with SqlCounter(query_count=2):
-        df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4]))
-        df.to_snowflake(test_table_name, if_exists="replace", index_label=[None])
-        verify_columns(test_table_name, ["index", "a", "b"])
+    if_exists = "replace"
+    df = pd.DataFrame({"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4]))
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists, index_label=[None])
+    verify_columns(test_table_name, ["index", "a", "b"])
 
 
 # one extra query to convert index to native pandas when creating the snowpark pandas dataframe
-@sql_count_checker(query_count=4)
 def test_to_snowflake_index_label_none_data_column_conflict(test_table_name):
     df = pd.DataFrame({"index": [1, 2, 3], "a": [4, 5, 6]})
-    df.to_snowflake(test_table_name, if_exists="replace")
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists)
     # If the column name "index" is taken by one of the data columns,
     # then "level_0" is used instead for naming the index column.
     # This is based on the behavior of reset_index.
     verify_columns(test_table_name, ["level_0", "index", "a"])
 
     df = pd.DataFrame(
-        {"index": [1, 2, 3], "a": [4, 5, 6]}, index=pd.Index([2, 3, 4], name="index")
+        {"index": [1, 2, 3], "a": [4, 5, 6]},
+        index=native_pd.Index([2, 3, 4], name="index"),
     )
     # If the index already has a name, "index", then "level_0" is not used,
     # and a ValueError is raised instead.
@@ -224,15 +317,16 @@ def test_to_snowflake_index_label_none_data_column_conflict(test_table_name):
 
     # nameless index
     df = pd.DataFrame({"index": [1, 2, 3], "a": [4, 5, 6]}, index=pd.Index([2, 3, 4]))
-    df.to_snowflake(test_table_name, if_exists="replace", index_label=[None])
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(test_table_name, if_exists=if_exists, index_label=[None])
     verify_columns(test_table_name, ["level_0", "index", "a"])
 
 
-# one extra query to convert index to native pandas when creating the snowpark pandas dataframe
-@sql_count_checker(query_count=1)
+@sql_count_checker(query_count=0)
 def test_to_snowflake_data_label_none_raises(test_table_name):
     df = pd.DataFrame(
-        {"a": [1, 2, 3], "b": [4, 5, 6]}, index=pd.Index([2, 3, 4], name="index")
+        {"a": [1, 2, 3], "b": [4, 5, 6]}, index=native_pd.Index([2, 3, 4], name="index")
     )
     df.columns = ["c", None]
 
@@ -243,11 +337,11 @@ def test_to_snowflake_data_label_none_raises(test_table_name):
         df.to_snowflake(test_table_name, if_exists="replace")
 
 
-@sql_count_checker(query_count=2)
-def test_to_snowflake_with_dropped_row_position():
+def test_to_snowflake_with_dropped_row_position(test_table_name):
     snow_df = pd.DataFrame({"a": [1, 2, 3], "b": [2, 4, 5]})
     snow_df = snow_df.groupby("a").count().reset_index()
-    snow_df.to_snowflake("out", index=False, table_type="temporary")
+    with to_snowflake_counter(df=snow_df, if_exists="fail"):
+        snow_df.to_snowflake(test_table_name, index=False, table_type="temporary")
 
 
 def verify_columns(table_name: str, expected: list[str]) -> None:
@@ -260,7 +354,6 @@ def verify_num_rows(session, table_name: str, expected: int) -> None:
     assert actual == expected
 
 
-@sql_count_checker(query_count=1)
 def test_timedelta_to_snowflake_with_read_snowflake(test_table_name, caplog):
     with caplog.at_level(logging.WARNING):
         df = pd.DataFrame(
@@ -270,7 +363,28 @@ def test_timedelta_to_snowflake_with_read_snowflake(test_table_name, caplog):
                 "t": native_pd.timedelta_range("1 days", periods=3),
             }
         )
-        df.to_snowflake(test_table_name, index=False, if_exists="replace")
+        if_exists = "replace"
+        with to_snowflake_counter(df=df, if_exists=if_exists):
+            df.to_snowflake(test_table_name, index=False, if_exists=if_exists)
         df = pd.read_snowflake(test_table_name)
-        assert df.dtypes[-1] == "int64"
+        assert df.dtypes[-1] == "float64"
         assert "`TimedeltaType` may be lost in `to_snowflake`'s result" in caplog.text
+
+
+@pytest.mark.parametrize(
+    "table_name_fixture",
+    (
+        "lower_case_table_name_unquoted",
+        "lower_case_table_name_quoted",
+        "upper_case_table_name_quoted",
+        "upper_case_table_name_with_space_quoted",
+    ),
+)
+def test_table_name(request, table_name_fixture):
+    name = request.getfixturevalue(table_name_fixture)
+    native_df = native_pd.DataFrame({"a": [1]})
+    df = pd.DataFrame(native_df)
+    if_exists = "replace"
+    with to_snowflake_counter(df=df, if_exists=if_exists):
+        df.to_snowflake(name, if_exists=if_exists, index=False)
+    pd.read_snowflake(name)